place on fragmentDefinition instead
