+++ .changeset/late-rabbits-burn.md
@@ -2,4 +2,5 @@
 'gql.tada': minor
 ---
 
-Add the possibility to unmask a fragment with a directive `@disable_mask`
+Add the possibility to unmask a fragment with a directive `@disable_mask` which can be
+added to a `FragmentDefinition`
